% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_grid.R
\name{access_grid}
\alias{access_grid}
\title{Access to gridded data following the MAS data structure.}
\usage{
access_grid(data_id, range = range, bbox = bbox)
}
\arguments{
\item{data_id}{Element of class \emph{character}.}

\item{range}{Two-element vector of class \emph{character} with desired start
and end dates (optional, in yyyy-mm-dd format).}

\item{bbox}{Four-element vector with target bounding box, composed of min. x,
min. y, max. x and max. y coordinates.}
}
\value{
A \emph{spatRaster} object
}
\description{
Access to gridded data following the MAS data structure.
}
\details{
{The function provides access to gridded data in the MAS database,
  returning as a \emph{SpatRaster} object, native to the \emph{terra} R
  package. If \emph{range} is provided, the function will select data files
  within the given temporal range and, if \emph{bbox is provided}, those
  files will be constrained to the given spatial bounding box. If the query
  encounters multiple files (i.e. time-series), the output will be a
  multi-layered \emph{spatRaster} object. Note that independently of the
  number of layers in the output, the function \link[terra]{time} can be used
  to determine the date to which the given layer corresponds to. However, if
  the target layer is a composite of multiple dates, this information will be
  set as NA. Whenever this function is ran, it will look for an existing
  temporary directory and store virtual rasters (VRT's) corresponding to the
  read-in layers. The temporary directory is created the first time this
  function is ran, and its path will be accessible using
  \emph{getOption('rast.temp')}.}
}
\examples{
{}
}
