% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_id.R
\name{build_id}
\alias{build_id}
\title{Build unique file identifier}
\usage{
build_id(dataset, subdataset, start_date, end_date, resolution)
}
\arguments{
\item{dataset}{Name of dataset (e.g. 'CCI_landCover').}

\item{subdataset}{Name of subdataset (e.g. 'landCover').}

\item{start_date}{Character element with start of dataset (as yyyy-mm-dd). When month/day is unknown, set to '00'.}

\item{end_date}{Character element with end of dataset (as yyyy-mm-dd). When month/day is unknown, set to '00'.}

\item{resolution}{Character element with resolution of dataset (e.g. '10km').}

\item{file}{Path to file, or files, to be registered in the database.}

\item{verbose}{Logical. Should the function report on the progress?}
}
\value{
Returns standardized file identifier.
}
\description{
Build unique file identifier
}
\details{
{Builds an unique file identifier, 
needed to register a file in the database.}
}
\seealso{
\code{\link{build_descriptor}}
}
