% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_project.R
\name{build_project}
\alias{build_project}
\title{Build project data structure}
\usage{
build_project(out.path)
}
\arguments{
\item{out.path}{Output directory where data structure will be built.}
}
\value{
A standardized file structure.
}
\description{
Build project data structure
}
\details{
{This function returns a standardized data structure to ease data and
code transferability. The final data structure includes the following folders:
\itemize{
\item{\emph{00_data -} Standardized data storage}
\item{\emph{01_analysis -} output of data analysis}
\item{\emph{02_documents -} General documents folder}
\item{\emph{03_code -} Storage of  project scripts}
\item{\emph{temp - temporary folder (used e.g. for testing)}}
}}
}
\examples{
{}

}
