% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sandbox.R
\name{build_sandbox}
\alias{build_sandbox}
\title{Create a sandbox containing small subsets of target datasets.}
\usage{
build_sandbox(data_id, bbox, range, path, virtual = FALSE, verbose = FALSE)
}
\arguments{
\item{data_id}{\emph{Character} vector with target dataset ID's.}

\item{bbox}{Four-element vector with target bounding box (min x, min y, max x, max y).}

\item{range}{Two-element \emph{Date} vector with start/end date range..}

\item{path}{\emph{Character} element with path to directory where to construct database.}

\item{virtual}{Logical. Should the output be a Virtual Raster (VRT) or a real image?}

\item{verbose}{Logical. Should the function communicate it's progress?}
}
\value{
A copy of a parent database.
}
\description{
Create a sandbox containing small subsets of target datasets.
}
\details{
{The function creates a copy of the datasets given by \emph{data_id}
within the area defined by \emph{bbox} and the temporal range determined by
\emph{range}. Currently, the function only handles raster objects, and it
provides the output either as a  \emph{GeoTiff}, if \emph{virtual} is FALSE,
or a \emph{VRT}, if \emph{virtual} is TRUE.}
}
