% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_vrt.R
\name{build_vrt}
\alias{build_vrt}
\title{Derives a VRT from a raster file}
\usage{
build_vrt(input, output, bbox = NULL)
}
\arguments{
\item{input}{Path to input raster file.}

\item{output}{Path to output VRT (optional).}

\item{bbox}{Four-element vector with target bounding box, composed of min. x, min. y, max. x and max. y coordinates.}
}
\value{
A \emph{VRT} file.
}
\description{
Derives a VRT from a raster file
}
\details{
{Creates a Virtual Raster (VRT) based on \emph{input}. When \emph{output} is missing, the function writes the
output to the default temporary folder. If \emph{bbox} is specified, it's used to constrain the spatial extent of the VRT.}
}
\examples{
{}

}
