% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_bibtex.R
\name{collect_bibtex}
\alias{collect_bibtex}
\title{Merge dataset references}
\usage{
collect_bibtex(data_id, output)
}
\arguments{
\item{data_id}{\emph{Character} vector enumerating datasets to query.}

\item{output}{Path to output file in bib format.}
}
\value{
A bibtex file combining references for requested datasets.
}
\description{
Merge dataset references
}
\details{
{This function provides access to the bibliographic 
references of one or more datasets in the MAS database. The 
function provides one bibtex file combining all relevant 
bibligraphic references. Note that literature references 
are only available at the top level. For example, the 
'CCI_landCover' dataset will be accepted in the function 
call, but 'CCI_landCover/landCover' will not return the 
desired output.}
}
