% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_metadata.R
\name{collect_metadata}
\alias{collect_metadata}
\title{MAS database builder.}
\usage{
collect_metadata(files)
}
\arguments{
\item{files}{\emph{Vector} listing files to evaluate}
}
\value{
RDS compiling relevant metadata for every file in the database
}
\description{
MAS database builder.
}
\details{
{Updates, or creates, RDS with metadata for every file in the 
database. the function will first search an existing file in the database, 
following the path set in \emph{getOptions('dmt.data')}. If no file exists, 
a new one will be created. If a file exists, the function will read it, 
list all files currently in the database, and update the list if the last 
modification date differs, or if new files were added.}
}
\seealso{
\code{\link{compile_metadata}}
}
