% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_dataset_metadata.R
\name{compile_dataset_metadata}
\alias{compile_dataset_metadata}
\title{Create metadata entry for the MAS database}
\usage{
compile_dataset_metadata(params = "params.yml")
}
\arguments{
\item{params}{\emph{Yaml} file with required metadata.}

\item{output}{output directory for derived metadata.}
}
\value{
A Rmarkdown file that can be fed into the MAS data catalog.
}
\description{
Create metadata entry for the MAS database
}
\details{
{The function parses a yaml file with metadata on a given dataset 
to create a standardized record of it. This record can then be integrated 
in the data catalog of the MAS research group. If no yaml is provided, the 
function will return a template as a variable, which can be saved with 
\link[yaml]{write_yaml}}.
}
