% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_metadata.R
\name{compile_metadata}
\alias{compile_metadata}
\title{MAS database builder.}
\usage{
compile_metadata(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical argument. Should the progress be reported?}
}
\value{
SQLite compiling relevant metadata for every file in the database
}
\description{
MAS database builder.
}
\details{
{Updates, or creates an SQLite file with metadata for every file
in the database. The function will first search for an existing database
file following the path set in \emph{getOption(dmt.data)}. If no file
exists, a new one will be created. If a file exists, the function will
read it, list all files currently in the database, and update the list
if the last modification date differs, or if new files were added.}
}
