% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Simple feature geometries of countries}
\format{
The object of class sf has 10 fields and 254 rows. It lists for each
  country the \code{unit} name, the \code{nation} name, \code{iso_a2} and
  \code{iso_a3}-code, whether the country is a \code{un_member} or an
  \code{island}, on which \code{continent} the country is and of which
  \code{region} and \code{subregion} that is part. Finally, it lists the
  initial administrative hiearchy ID (ahID) for each nation and contains a
  geometry for each feature.
}
\usage{
countries
}
\description{
Simple feature geometries of countries
}
\keyword{datasets}
