% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covariates}
\alias{covariates}
\title{Target covariates}
\format{
The tibble contains the columns 'product', 'variable', 'resolution'
  and 'time' and describes gridded data products that were harmonised by
  Ruben Remelgado to a common resolution.
}
\usage{
covariates
}
\description{
Target covariates
}
\keyword{datasets}
