% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmt_tmp.R
\name{dmt_tmp}
\alias{dmt_tmp}
\title{Retrieves or creates a session-specific temporary directory.}
\usage{
dmt_tmp(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. Should the function inform on progress? Default is FALSE.}
}
\value{
A \emph{character} element.
}
\description{
Retrieves or creates a session-specific temporary directory.
}
\details{
The function identifies/creates the DMT temporary directory, which is required to handle raster data.
}
\examples{
{}

}
