% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{documentation}
\alias{documentation}
\title{Access documentation websites}
\usage{
documentation(source)
}
\arguments{
\item{source}{What documentation to access? 'Choose `catalog` or `toolbox`.}
}
\description{
Access documentation websites
}
\details{
{Helper tool to access documentation websites of the MAS database. 
If the user chooses `catalog`, the function will open the website for the 
data catalog. If the user chooses `toolbox`, the function will open the 
website for the masDMT R package.}
}
