% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_database.R
\name{.access_database}
\alias{.access_database}
\title{Tool to open database file}
\usage{
.access_database(database)
}
\arguments{
\item{database}{Optional. data.frame with metadata.}
}
\value{
a data.frame, if \emph{database} is provided.
}
\description{
Tool to open database file
}
\details{
{Returns an existing database registry as a data.frame if
\emph{database} is missing, the function updates the registry.}
}
