% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMeta.R
\name{.build_meta}
\alias{.build_meta}
\title{MAS database builder.}
\usage{
.build_meta(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical argument. Should the progress be reported? Default is FALSE.}
}
\value{
An SQLite database
}
\description{
MAS database builder.
}
\details{
Updates, or creates, SQL database with metadata on MAS data products.
}
\examples{
{}

}
\seealso{
\code{\link{access_grid}} \code{\link{access_vector}} \code{\link{list_data}}
}
