% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock_database.R
\name{.lock_database}
\alias{.lock_database}
\title{Lock database to allow for unobstructed reading/writing}
\usage{
.lock_database()
}
\value{
A lock file
}
\description{
Lock database to allow for unobstructed reading/writing
}
\details{
Creates a file lock that avoids changes to a datavase registry.
}
