% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legend.R
\name{get_legend}
\alias{get_legend}
\title{Accesses legend for a given product.}
\usage{
get_legend(data.id)
}
\arguments{
\item{data.id}{Element or vector of class \emph{character}.}
}
\value{
{Returns a |emph{data.frame} with the legend of the target dataset, containing:
\itemize{
\item{\emph{value} - Grid value}
\item{\emph{label} - class label}
\item{\emph{hex} - Hex color code}
}}
}
\description{
Accesses legend for a given product.
}
\details{
The function provides access to legend information relative to data products existing in the MAS database.
}
