% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_descriptor.R
\name{knit_descriptor}
\alias{knit_descriptor}
\title{Knits a metadata entry for the MAS database}
\usage{
knit_descriptor(params)
}
\arguments{
\item{params}{\emph{Yaml} file with required metadata.}
}
\value{
A Rmarkdown file that can be fed into the MAS data catalog.
}
\description{
Knits a metadata entry for the MAS database
}
\details{
{The function parses a yaml file with metadata on a given
dataset to create a standardized, html record of it. This record can
then be added to the data catalog of the MAS research group.}
}
