% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockpick.R
\name{lockpick}
\alias{lockpick}
\title{Change file permissions}
\usage{
lockpick(path, mode)
}
\arguments{
\item{path}{Path to the file for which to change permissions.}

\item{mode}{Character element. One of 'close' or 'open'.}
}
\value{
File permission changes.
}
\description{
Change file permissions
}
\details{
{This function will change the target file permissions to protect
it from accidental deletion and changes. When \emph{mode} is set to "close",
the function will apply a chmod code "444", making the file "read-only".
If set to "open", the function will apply the chmod code "666", enabling
changes to the file, including deletion. This will prevent other users from modifying
the file, but will allow the user to If this function is ran in windows,
the function will additionally use
\href{https://docs.microsoft.com/en-us/windows-server/administration/windows-commands/icacls}{
icacls to modify file permissions}. Since windows cannot interpret the chmod codes.}
}
