% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly2sample.R
\name{poly2sample}
\alias{poly2sample}
\title{Converts polygons into points and reports on their raster overlap}
\usage{
poly2sample(x, y, id)
}
\arguments{
\item{x}{Object of class \emph{SpatVector} and of type \emph{polygon}.}

\item{y}{Single or two-element vector of class \emph{numeric} with the target resolution.}

\item{id}{\emph{Numeric} or \emph{character} vector with unique identifiers, one for each row of \emph{x}.}
}
\value{
{A \emph{data.frame} object, containing:
\itemize{
\item{\emph{id}: Unique identifier of the reference polygon}
\item{\emph{x}: x coordinate}
\item{\emph{y}: y coordinate}
\item{\emph{overlap}: percent overlap between the pixel and the polygon}
}}
}
\description{
Converts polygons into points and reports on their raster overlap
}
\details{
{The function iterates through each element of \emph{x}, converting the corresponding polygon into points. For
each point, the function returns an estimate of the percent overlap between the corresponding pixel and the reference
polygon. Each entry in the output \emph{data.frame} will be related to the original reference polygon through a unique
identifier. Without an \emph{id} argument, the identifier will reflect the order by which \emph{x} was sorted.}
}
\examples{
{}

}
\keyword{sampling}
