% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_arealData.R
\name{pull_arealData}
\alias{pull_arealData}
\title{Pull the areal data from an arealDB database}
\usage{
pull_arealData(
  path = NULL,
  nation = NULL,
  pID = NULL,
  geoID = NULL,
  tabID = NULL,
  type = NULL,
  commodities = NULL,
  keep_ids = TRUE
)
}
\arguments{
\item{path}{[\code{character(1)}]\cr the path of the database.}

\item{nation}{[\code{character(1)}]\cr the nation for which to get the data.}

\item{pID}{[\code{character(1)}]\cr name of the primary ID.}

\item{geoID}{[\code{itegerish(1)}]\cr subset the data by geoID.}

\item{tabID}{[\code{itegerish(1)}]\cr subset the data by tabID.}

\item{type}{[\code{character(1)}]\cr the target variable to set as demand
type.}

\item{commodities}{[\code{character(.)}]\cr the target commodities; all other
commodities are grouped into the group "other".}

\item{keep_ids}{[\code{logical(1)}]\cr}
}
\description{
Pull the areal data from an arealDB database
}
