% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_geometries.R
\name{pull_geometries}
\alias{pull_geometries}
\title{Pull the geometries from an arealDB database}
\usage{
pull_geometries(path = NULL, nation = NULL, layer = NULL, geoID = NULL)
}
\arguments{
\item{path}{[\code{character(1)}]\cr the path of the database.}

\item{nation}{[\code{character(1)}]\cr the nation for which to get the data}

\item{layer}{[\code{character(1)}]\cr the name of the administrative level to
get.}

\item{geoID}{[\code{itegerish(1)}]\cr subset the data by geoID.}
}
\description{
Pull the geometries from an arealDB database
}
