% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_ontology.R
\name{pull_ontology}
\alias{pull_ontology}
\title{Pull an ontology}
\usage{
pull_ontology(path = NULL)
}
\arguments{
\item{path}{[\code{character(1)}]\cr the path of the ontology}
}
\value{
a tibble that contains at least the columns \code{term_from},
  \code{class_from}, \code{relationship}, \code{term_to} and \code{class_to}
}
\description{
Pull an ontology
}
\details{
This funciton expects a csv document that contains a tabular
  ontology, i.e., a table that contains three kinds of information, the
  "from" part, the "relationship" and the "to" part. Both, from and to parts
  are a combination of the term an its class, which results in at least 5
  columns in total, but could also include other columns, such as additional
  IDs or quality metrics for each relationship.
}
