% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_legend.R
\name{register_legend}
\alias{register_legend}
\title{Register legend file}
\usage{
register_legend(data_id, label, value, r, g, b, overwrite = FALSE)
}
\arguments{
\item{data_id}{Dataset id, composed of \emph{dataset} and \emph{subdataset} (e.g. 'CCI_landCover/landCover').}

\item{label}{Character vector with labels matching input of \emph{value}.}

\item{value}{Numeric vector with unique grid value.}

\item{r}{Numeric vector with red color intensity (0-255).}

\item{g}{Numeric vector with green color intensity (0-255).}

\item{b}{Numeric vector with blue color intensity (0-255).}

\item{overwrite}{Logical. If a legend already exists, should it be modified?}
}
\value{
Writes file into standardized file structure.
}
\description{
Register legend file
}
\details{
{Registers a the legend for a gridded dataset,
which can be used for standardized mapping and plotting.}
}
\seealso{
\code{\link{build_descriptor}}
}
