% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_path.R
\name{register_path}
\alias{register_path}
\title{Register relevant paths}
\usage{
register_path(paths)
}
\arguments{
\item{paths}{List of parameters.}
}
\value{
Writes file into standardized file structure.
}
\description{
Register relevant paths
}
\details{
{Registers paths of different material of a project. By default,
the function accepts a the following keywords, provided in a named list:
\itemize{
\item{"dmt.data"}{Path to database (default set to iDiv's MAS drive)}
\item{"dmt.catalog"}{Path to data catalog (default set to
\href{https://macroecology-society.github.io/data-catalog/}{MAS data catalog})}
\item{"dmt.toolbox"}{Path to the documentation of a toolbox for a project}
}
The provided paths will be recorded in a configuration file within
\emph{masDMT}, meaning they will set as defaults in future R sessions.}
}
\seealso{
\code{\link{list_data}} \code{\link{documentation}}
}
