% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_ontology.R
\name{summarise_ontology}
\alias{summarise_ontology}
\title{Print a summary of an ontology}
\usage{
summarise_ontology(ontology = NULL, names_from = "class_to")
}
\arguments{
\item{ontology}{[\code{data.frame(1)}]\cr the ontology to summarise}
}
\description{
Print a summary of an ontology
}
\details{
This function gives a summary of an ontology by listing all unique
  items in the column \code{term_from} and counting how many and which
  relationships to all \code{class_to} are defined for each term.
}
