% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_indices.R
\name{table_indices}
\alias{table_indices}
\title{Returns indices to process large tables}
\usage{
table_indices(path, size)
}
\arguments{
\item{path}{Path to tabular data file.}

\item{size}{Numeric element specifying the max. data volume (in Mb).}
}
\value{
\emph{list} object with indices
}
\description{
Returns indices to process large tables
}
\details{
{This function helps process tables that are larger
than memory. Given a file \emph{path}, the function infers the
storage size of the input file and determines an optimal number
of runs. To do, the argument \emph{size} is used, which determines
the maximum amount of RAM that should be occupied in a single read.
Then, for each of the n runs, the function determines the corresponding
start and end rows, which are reported as a list. The output can be used
in combination with, e.g., \code{\link[readr]{read_delim}}, which allows
users to specify starting and ending rows when importing a file.}
}
