% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_descriptor.R
\name{build_descriptor}
\alias{build_descriptor}
\title{Creates a metadata record for the data catalog}
\usage{
build_descriptor(path, data_id, bibtex = NULL, preview = NULL)
}
\arguments{
\item{path}{Directory where to write metadata record.}

\item{data_id}{Dataset unique identifier.}

\item{bibtex}{Path to bibtex with dataset references. Optional.}

\item{preview}{Path to preview image.}
}
\value{
A Rmarkdown file that can be fed into the MAS data catalog. Optional.
}
\description{
Creates a metadata record for the data catalog
}
\details{
{The function creates a directory in the specified \emph{path}
named as the \emph{data_id}. Within it, you will find a yaml file where
you can manually edit the metadata for the new record. If a \emph{bibtex}
and/or a \emph{preview} image are provided, the function will copy them
into the new directory after testing their validity. Once filled, the
yaml file can be compiled into a metadata record using
\code{\link[masDMT]{compile_metadata}}.}
}
