% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_descriptor.R
\name{build_descriptor}
\alias{build_descriptor}
\title{Create metadata entry for the MAS database}
\usage{
build_descriptor(params = "params.yml")
}
\arguments{
\item{params}{\emph{Yaml} file with required metadata.}
}
\value{
A Rmarkdown file that can be fed into the MAS data catalog.
}
\description{
Create metadata entry for the MAS database
}
\details{
{The function parses a yaml file with metadata on a given dataset 
to create a standardized record of it. This record can then be integrated 
in the data catalog of the MAS research group. If no yaml is provided, the 
function will write a template in the current working directory,}
}
