% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_id.R
\name{build_id}
\alias{build_id}
\title{Build unique file identifier}
\usage{
build_id(dataset, subdataset, start_date, end_date, resolution)
}
\arguments{
\item{dataset}{Name of dataset (e.g. 'CCI_landCover').}

\item{subdataset}{Name of subdataset (e.g. 'landCover').}

\item{start_date}{Date object with start date of the dataset.}

\item{end_date}{Date object with end date of the dataset.}

\item{resolution}{Character element with resolution of dataset (e.g. '10km').}
}
\value{
Returns standardized file identifier.
}
\description{
Build unique file identifier
}
\details{
{Builds an unique file identifier,
needed to register a file in the database.}
}
\seealso{
\code{\link{register_data}}
}
