% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_data.R
\name{register_data}
\alias{register_data}
\title{register dataset}
\usage{
register_data(
  dataset,
  subdataset,
  start_date,
  end_date,
  resolution,
  catalog = TRUE,
  replace = TRUE,
  delete = FALSE,
  verbose = TRUE,
  file = NULL
)
}
\arguments{
\item{dataset}{Name of dataset (e.g. 'CCI_landCover').}

\item{subdataset}{Name of subdataset (e.g. 'landCover').}

\item{start_date}{Character element with end of dataset (as yyyy-mm-dd). When month/day is unknown, set to '00'.}

\item{resolution}{Character element with resolution of dataset (e.g. '10km').}

\item{catalog}{Logical. Should a new metadata entry be added to the database?}

\item{replace}{Logical. Replace file in catalog if the name is repeated?}

\item{delete}{Logical. Delete original file when finished?}

\item{verbose}{Logical. Should the function report on the progress?}

\item{file}{Path to file, or files, to be registered in the database.}
}
\value{
Writes file into standardized file structure.
}
\description{
register dataset
}
\details{
{When file is set to NULL, the function will return a standardized 
file path, without extension, where the file would need to be stored. If 
\emph{file} is not NULL, then the function will search for the specified 
file in system, and make a copy of it in the data structure using the 
standardized name. The fill will in addition be registered in the existing 
flat database, given that \emph{catalog} is set to TRUE.}
}
\seealso{
\code{\link{build_descriptor}}
}
